有模型颜色了。 原因：shader中的Indices 类型与AS、hlslCompat中的不同

+++ D3D12RaytracingSimpleLighting/RaytracingHlslCompat.h
@@ -28,7 +28,7 @@ using namespace DirectX;
 
 
 // Shader will use byte encoding to access indices.
-typedef UINT16 Index;
+typedef uint32_t Index;
 #endif
 
 struct Mesh
